backend-b-intelligence/requirements.txt
flask==3.0.0
openai==1.12.0
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
requests==2.31.0
httpx<0.28
